% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucminf-package.R
\name{ucminf-package}
\alias{ucminf-package}
\alias{_UCMINF}
\title{ucminf: General-Purpose Unconstrained Non-Linear Optimization}
\description{
The \pkg{ucminf} package provides an algorithm for general-purpose
unconstrained non-linear optimization.
}
\section{Bugreport}{
 Any bug or suggestion can be reported using the
\code{ucminf} tracker facilities at:
\url{https://github.com/hdakpo/ucminf/issues}
}

\author{
K Hervé Dakpo, Hans Bruun Nielsen, and Stig Bousgaard Mortensen
}
